+++ LAB_12_13_14/src/PI_Controller/PI_controller.h
@@ -1,16 +1,16 @@
-/* 
- * File:   PI_controller.h
- * Author: Andre Brandao
- * Author: Emanuel Pereira
- *
- *
+/** \file PI_controller.h
+ * 	\brief Module implementing PI Controller
+ * 
+ * \author André Brandão
+ * \author Emanuel Pereira
+ * \date 15/06/2022
  */
+ 
 
 #ifndef _PI_CONTROLLER_H
 #define	_PI_CONTROLLER_H
 
-
-void PI_init(int, int);
-double PI_controller(int, int, double);
+void PI_init(float, float);
+int PI_controller(int, int, int);
 
 #endif	// _PI_CONTROLLER_H
